cloudrun_api/requirements.txt
Flask==2.3.3
gunicorn==20.1.0
google-cloud-storage==2.11.0
requests==2.32.3
google-cloud-aiplatform==1.63.0
